+++ src/main/java/chess/domain/pieces/Rook.java
@@ -1,5 +1,8 @@
 package chess.domain.pieces;
 
+import chess.domain.position.Position;
+import chess.domain.position.Row;
+
 public class Rook extends Piece {
     private static final String BLACK_TEAM_ROW = "8";
     private static final String WHITE_TEAM_ROW = "1";
