+++ src/main/java/chess/domain/position/Row.java
@@ -1,4 +1,4 @@
-package chess.domain.pieces;
+package chess.domain.position;
 
 import java.util.Arrays;
