+++ src/test/java/chess/domain/position/ColTest.java
@@ -1,5 +1,6 @@
-package chess.domain.pieces;
+package chess.domain.position;
 
+import chess.domain.position.Col;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
