+++ src/test/java/chess/domain/position/PositionTest.java
@@ -1,4 +1,4 @@
-package chess.domain.pieces;
+package chess.domain.position;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
